[CL-2362] Move topics service
