+++ front/app/services/topics.ts
@@ -67,3 +67,34 @@ export function topicByIdStream(topicId: string) {
 export function topicsStream(streamParams: ITopicsStreamParams | null = null) {
   return streams.get<ITopics>({ apiEndpoint, ...streamParams });
 }
+
+export interface ITopicUpdate {
+  title_multiloc: Multiloc;
+  description_multiloc: Multiloc;
+}
+
+export async function addTopic(object: ITopicUpdate) {
+  const response = await streams.add<ITopic>(apiEndpoint, { topic: object });
+  await streams.fetchAllWith({ apiEndpoint: [apiEndpoint] });
+  return response;
+}
+
+export async function updateTopic(topicId: string, object: ITopicUpdate) {
+  const response = await streams.update<ITopic>(
+    `${apiEndpoint}/${topicId}`,
+    topicId,
+    {
+      topic: object,
+    }
+  );
+
+  await streams.fetchAllWith({ apiEndpoint: [apiEndpoint] });
+  return response;
+}
+
+export async function deleteTopic(topicId: string) {
+  const response = await streams.delete(`${apiEndpoint}/${topicId}`, topicId);
+  await streams.fetchAllWith({ apiEndpoint: [apiEndpoint] });
+
+  return response;
+}
